adding some essential codes and browser gets hello arya

+++ include/tcp.h
@@ -1,8 +1,8 @@
 #ifndef _TCP_H
 #define _TCP_H
 
-int tcpsock_bind(char * host, int port, int backlog);
-int tcpsock_connect(char * host, int port);
+int tcpsock_bind(char * host,unsigned int port,unsigned short backlog);
+int tcpsock_connect(char * host,unsigned int port);
 int tcpsock_accept(int sfd);
 
 #endif // _TCP_H
